requirements.txt
# Core ML and FL
tensorflow==2.13.0
flwr==1.5.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# PSI
openmined.psi==1.0.1

# API
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
python-multipart==0.0.6

# Security and Crypto
cryptography==41.0.3

# Database and Cache
redis==5.0.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.20

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1

# Utilities
python-dotenv==1.0.0